services/ocr/requirements.txt
fastapi
uvicorn[standard]
Pillow
numpy
torch          # your chosen version (CPU or CUDA)
transformers>=4.52.0
accelerate>=0.34.0
